+++ fltk/examples/help_view.rs
@@ -1,0 +1,11 @@
+use fltk::*;
+
+fn main() {
+    let app = app::App::default();
+    let mut win = window::Window::new(100, 100, 400, 300, "Help View");
+    let mut h = misc::HelpView::new(10, 10, 380, 280, "");
+    h.set_value("Hello <b><font color=red>again</font></b>");
+    win.end();
+    win.show();
+    app.run().unwrap();
+}